pkg/helm/installer: resource reconciliation clarity improvements
* new functions for resource reconciliation
* moved New* functions to separate file
